+++ Cargo.toml
@@ -29,8 +29,13 @@ atuin = { path = "/usr/bin/atuin" }
 [workspace]
 members = ["./atuin-client", "./atuin-server", "./atuin-common"]
 
+[features]
+default = ["sync", "server"]
+sync = ["atuin-client/sync"]
+server = ["atuin-server"]
+
 [dependencies]
-atuin-server = { path = "atuin-server", version = "0.8.1" }
+atuin-server = { path = "atuin-server", version = "0.8.1", optional = true }
 atuin-client = { path = "atuin-client", version = "0.8.1" }
 atuin-common = { path = "atuin-common", version = "0.8.1" }
