+++ atuin-client/src/lib.rs
@@ -3,11 +3,15 @@
 #[macro_use]
 extern crate log;
 
+#[cfg(feature = "sync")]
 pub mod api_client;
-pub mod database;
+#[cfg(feature = "sync")]
+pub mod sync;
+#[cfg(feature = "sync")]
 pub mod encryption;
+
+pub mod database;
 pub mod history;
 pub mod import;
 pub mod ordering;
 pub mod settings;
-pub mod sync;
