+++ src/command/client/history.rs
@@ -9,6 +9,8 @@ use tabwriter::TabWriter;
 use atuin_client::database::Database;
 use atuin_client::history::History;
 use atuin_client::settings::Settings;
+
+#[cfg(feature = "sync")]
 use atuin_client::sync;
 
 #[derive(Subcommand)]
@@ -141,8 +143,12 @@ impl Cmd {
                 db.update(&h).await?;
 
                 if settings.should_sync()? {
-                    debug!("running periodic background sync");
-                    sync::sync(settings, false, db).await?;
+                    #[cfg(feature = "sync")] {
+                        debug!("running periodic background sync");
+                        sync::sync(settings, false, db).await?;
+                    }
+                    #[cfg(not(feature = "sync"))]
+                    debug!("not compiled with sync support");
                 } else {
                     debug!("sync disabled! not syncing");
                 }
