+++ src/command/client/sync/logout.rs
