+++ src/command/client/sync/register.rs
