+++ src/command/mod.rs
@@ -2,6 +2,8 @@ use clap::Subcommand;
 use eyre::Result;
 
 mod client;
+
+#[cfg(feature = "server")]
 mod server;
 
 #[derive(Subcommand)]
@@ -11,6 +13,7 @@ pub enum AtuinCmd {
     Client(client::Cmd),
 
     /// Start an atuin server
+    #[cfg(feature = "server")]
     #[clap(subcommand)]
     Server(server::Cmd),
 }
@@ -19,6 +22,7 @@ impl AtuinCmd {
     pub async fn run(self) -> Result<()> {
         match self {
             Self::Client(client) => client.run().await,
+            #[cfg(feature = "server")]
             Self::Server(server) => server.run().await,
         }
     }
